Filterfunction intertwined with the price bar / renamed the studio.js to studios.js / erased garbage-code
