--- src/api/studio.js
@@ -1,344 +1,0 @@
-export const studios = [
-  {
-    // Logo Größe 250 250px
-    imageLogo:
-      "https://www.fitnessfirst.de/modules/custom/ff_custom/images/logo.png",
-    imageSource:
-      "https://media.cylex.de/companies/7027/998/images/-1174487329-Fitness-First-K%C3%B6ln-Ehrenfeld-Freestyle-Fl%C3%A4che_500999_large.jpg",
-    name: "Fitness First Ehrenfeld",
-    contracts: ["24 Months  40€ p.Month"],
-    price: [40],
-    duration: [" 24 "],
-    courses: ["Zumba  ", "Yoga", "Body-Pump"],
-    wellnes: ["Spa", "Sauna"],
-    cardio: ["Rowing", "Stepper"],
-    strength: ["Hanteln", "Geräte"],
-
-    neighbourhood: "Ehrenfeld",
-    adress: "Venloer Str. 389 50825 Cologne",
-    website: "https://www.fitnessfirst.de/clubs/koeln-ehrenfeld",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2513.537297175328!2d6.913156715749572!3d50.950769579547604!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf253932ab8b01%3A0xeddd7d742fd472e3!2sVenloer%20Str.%20389%2C%2050825%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572284671677!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: false,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: false,
-      six: false,
-      twelve: false,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    imageLogo:
-      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSTdXhCx0Liy27EUNHcXI5T6YbRC5E-6-0IRI52ANRk-nvwyfCyDQ&s",
-    imageSource:
-      "https://intpass-eu-central-1-live02-public.s3.eu-central-1.amazonaws.com/de-live/venueHome_1024x576_KOE_156509059957126_admin.jpg",
-    name: "FitnessLoft",
-    contracts: ["flex  80€ p.Month ", "24 Months  30€ p.Month"],
-    price: [80, 30],
-    duration: [" flex ", " 24  "],
-    courses: [
-      "Yoga",
-      "EDM",
-      "Bodypump",
-      "Bodyshape",
-      "Boxen",
-      "Bauch Beine Po"
-    ],
-    wellnes: ["Sauna"],
-    cardio: ["Laufband", "Stepper", "Crosstrainer", "Fahrad-Ergometer"],
-    strength: ["Kraftgeräte", "Hantelbereich"],
-
-    neighbourhood: ["Altstadt-Nord"],
-    adress: "Eigelstein 80-88 50668 Cologne",
-    website: "https://fitnessloft.de/fitnessstudios/koeln/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2513.6703135834955!2d6.955026715749513!3d50.94830987954713!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf25a31a4fb843%3A0x9d9f19559be08f16!2sEigelstein%2080-88%2C%2050668%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572284825454!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: false,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: true,
-      six: true,
-      twelve: true,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    imageLogo:
-      "https://upload.wikimedia.org/wikipedia/commons/6/6b/McFit_Logo.png",
-    imageSource:
-      "https://www.mcfit.com/uploads/pics/McFIT_Studiobilder_1200x500_DE_12a.jpg",
-    name: "Mc Fit Altstadt",
-    contracts: ["24 Months  19,90€ p.Month"],
-    price: [19.9],
-    duration: ["24"],
-
-    courses: ["CyberKurse", "Crossworkout", "Pre-Running", "Spinning"],
-    wellnes: [],
-    cardio: ["Stepper", "Rudern", "Crosstrainer"],
-    strength: ["Hantelbereich", "Kraftgeräte", "Zirkeltraining"],
-
-    neighbourhood: ["Innenstadt"],
-    adress: "Hohe Str. 62-68 50667 Cologne",
-    website:
-      "https://www.mcfit.com/de/studios/studiosuche/studiodetails/studio/koeln-altstadt/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2514.3020958260017!2d6.9543310157490685!3d50.9366259795454!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf25ae1d3eaf7d%3A0x2e387755eea44033!2sHohe%20Str.%2062%2C%2050667%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572284928497!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: true,
-      personaltraining: false,
-      cardio: true,
-      wellnes: false,
-      strength: true,
-      courses: true,
-      flexx: false,
-      six: false,
-      twelve: false,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    imageLogo:
-      "https://better-side.com/wp-content/uploads/2019/04/16649228_10155076498164791_1462676470057436642_n.png",
-    imageSource:
-      "https://studiosuche.de/assets/images/chains/holmes-place/holmes-place-circle.jpg",
-    name: "Holmes Place am Gürzenich",
-    contracts: [
-      "flexx  89,90€ p.Month",
-      "12 Months  74,90 p.Month ",
-      "24 Months  64,90€ p.Month"
-    ],
-    price: [89.9, 74.9, 64.9],
-
-    duration: ["flex", "12", "24"],
-    courses: ["Yoga", "Bodypump", "Aerobic", "TaiChi", "Zumba", "Bodycombat"],
-    wellnes: ["Pool", "Sauna", "Spa", "Solarium", "Massage"],
-    cardio: [
-      "Laufband",
-      "Stepper",
-      "Crosstrainer",
-      "Fahrad-Ergometer",
-      "Spinning-Bike",
-      "Ruder-Ergometer"
-    ],
-    strength: ["Hantelbereich", "Zirkeltraining", "Kraftgeräte"],
-
-    neighbourhood: ["Altstadt-Nord", "Innenstadt"],
-    adress: "Gürzenichstraße 6 50667 Cologne",
-    website: "http://holmesplace.de/am-guerzenich.c/de/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2514.331000691899!2d6.955005515749042!3d50.936091379545346!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf25b1fb85118b%3A0xfedd06dafc5dfef5!2zR8O8cnplbmljaHN0cmHDn2UgNiwgNTA2NjcgS8O2bG4!5e0!3m2!1sde!2sde!4v1572285025806!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: true,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: true,
-      six: true,
-      twelve: true,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    //Logo Größe 180 x180
-    imageLogo:
-      "https://cdn.11880.com/fitness-studio-bushido-fitnessstudio_17345142_mw640h480_koeln.jpg",
-    imageSource:
-      "https://files.eversports.com/e6663d71-42c0-4f0c-98e2-a559a0c4f885/mg_0003-1280jpg-x-large.jpg",
-    name: "Bushido Fitness",
-    contracts: [
-      "6 Months  58€ p.Month",
-      "12 Months  53€ p.Month ",
-      "24 Months  48€ p.Month"
-    ],
-    price: [58, 53, 48],
-    duration: [" 6 ", " 12 ", " 24 "],
-    courses: ["Bodypup", "Yin-Yoga", "Pilates", "RehaRücken"],
-    wellnes: ["Entspannungsraum", "Massagen", "Sauna"],
-    cardio: ["Laufband", "Ruder-Ergometer", "Fahrradergometer"],
-    strength: ["Hantelbereich", "Kraftgeräte", "Zirkeltraining"],
-
-    neighbourhood: ["Innenstadt"],
-    adress: "Mittelstraße 12, 50672 Cologne",
-    website: "https://www.bushido-fitness.de/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2514.279554636692!2d6.94186621574907!3d50.9370428795455!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf250755f81cdf%3A0x795ba63b57741430!2sMittelstra%C3%9Fe%2012%2C%2050672%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572285216985!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: false,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: false,
-      six: true,
-      twelve: true,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    imageLogo: "http://kides.de/wp-content/uploads/2014/11/art-of-fitness.jpg",
-    imageSource:
-      "https://files.eversports.com/a8d82428-a4e2-4e6b-9e50-c1c5b3098e84/art-of-fitness-1-x-large.jpg",
-    name: "Art of Fitness",
-    contracts: [
-      "flexx  39,90€ p.Month",
-      "6 Months  29,90€ p.Month ",
-      "12 Months 24,90€ p.Month"
-    ],
-    price: [39.9, 29.9, 24.9],
-    duration: ["flex", "6", "12"],
-    courses: ["Yoga", "Bodypump", "Aerobic", "TaiChi", "Zumba", "Bodycombat"],
-    wellnes: ["Sauna", "Solarium", "Spa"],
-    cardio: [
-      "Laufband",
-      "Stepper",
-      "Crosstrainer",
-      "Fahrad-Ergometer",
-      "Spinning-Bike",
-      "Ruder-Ergometer"
-    ],
-    strength: ["Hantelbereich", "Kraftgeräte", "Zirkeltraining"],
-
-    neighbourhood: ["Ehrenfeld"],
-    adress: "Maarweg 149-152 50825 Cologne",
-    website: "http://art-of-fitness.de/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2513.874796252627!2d6.8956554157493715!3d50.944528479546605!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf252d8aaffe67%3A0xaa5ebea5625f6456!2sMaarweg%20149%2C%2050825%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572285304898!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: false,
-      personaltraining: true,
-      cardio: false,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: true,
-      six: true,
-      twelve: true,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-
-  {
-    imageLogo:
-      "https://www.fitnessfirst.de/modules/custom/ff_custom/images/logo.png",
-    imageSource: "src",
-    name: "Fittness First-Innenstadt",
-    contracts: ["24 Months  40€ p.Month"],
-    price: [40],
-    duration: ["24"],
-
-    courses: [
-      "Zumba",
-      "Yoga",
-      "Body-Pump",
-      "Tae Bo",
-      "Energy Cycling",
-      "Cardio-Step-Training"
-    ],
-
-    wellnes: ["Spa", "Sauna", "Solarioum", "Massagen "],
-    cardio: [
-      "Laufband",
-      "Stepper",
-      "Crosstrainer",
-      "Fahrad-Ergometer",
-      "Spinning-Bike",
-      "Ruder-Ergometer"
-    ],
-    strength: ["Hantelbereich", "Kraftgeräte"],
-
-    neighbourhood: ["Innenstadt"],
-    adress: "Schildergasse 94 50667 Cologne",
-    website: "http://www.fitnessfirst.de/clubs/koeln-innenstadt-schildergasse/",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2514.2923158893377!2d6.94911911563744!3d50.93680686008173!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf25aea5ee92e5%3A0xd23983ebc746d742!2sZara%2C%20Schildergasse%2094%2C%2050667%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572285420909!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: true,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: false,
-      six: false,
-      twelve: false,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  },
-  {
-    imageLogo:
-      "https://cdn.pixadus.com/2ecd4ce6-def1-4c2f-8858-5401504d8c81.png",
-    imageSource: "src",
-    name: "Just Fit 13",
-    contracts: [
-      "flexx  75€ p.Month",
-      "12 Months  55€ p.Month ",
-      "24 Months  45€ p.Month"
-    ],
-    price: [75, 55, 45],
-    duration: ["flex", "12", "24"],
-    courses: [
-      "Yoga",
-      "Body-Pump",
-      "Tae Bo",
-      "Energy Cycling",
-      "Cardio-Step-Training"
-    ],
-    wellnes: ["Spa", "Sauna", "Solarioum", "Massagen "],
-    cardio: [
-      "Laufband",
-      "Stepper",
-      "Crosstrainer",
-      "Fahrad-Ergometer",
-      "Spinning-Bike",
-      "Ruder-Ergometer"
-    ],
-    strength: ["Hantelbereich", "Kraftgeräte", "Zirkeltraining"],
-
-    neighbourhood: ["Altstadt-Süd"],
-    adress: "Hohenstaufenring 30-32, 50674 Cologne",
-    website:
-      "https://www.justfit-clubs.de/just-fit-13-city-koeln-altstadt-sued.html",
-    iFrameSource:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2514.574217870724!2d6.938767315748905!3d50.93159287954459!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47bf24ffe1f74ca3%3A0x89b12f6ab186794c!2sHohenstaufenring%2030%2C%2050674%20K%C3%B6ln!5e0!3m2!1sde!2sde!4v1572285652669!5m2!1sde!2sde",
-    city: "Cologne",
-    options: {
-      ladyarea: false,
-      personaltraining: true,
-      cardio: true,
-      wellnes: true,
-      strength: true,
-      courses: true,
-      flexx: true,
-      six: true,
-      twelve: true,
-      twentyfour: true
-    },
-    openingtime: "Mo-Th: 7-23 / Fr: 7-22 / Sa-Su: 10-22"
-  }
-];
